pyproject.toml
[project]
name = "trade-dash"
version = "0.1.0"
description = "Trading dashboard application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.56.0",
]

[project.scripts]
trade-dash = "trade_dash.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trade_dash"]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pandas-stubs>=3.0.0.260204",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
    "SIM",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
    "--cov=trade_dash",
    "--cov-report=term-missing",
    "-ra",
]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["trade_dash"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.mypy]
python_version = "3.11"
files = ["src", "tests"]
strict = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
